requirements.txt
opencv-python~=4.12.0.88
mediapipe~=0.10.14
pywin32
pystray~=0.19.5
Pillow~=11.3.0

customtkinter~=5.2.2